Fixed bug causing team pages to load slower
