+++ src/layouts/index.tsx
@@ -1,4 +1,3 @@
 export { default as DialogLayout } from "./DialogLayout"
 export { default as AuthenticatedLayout } from "./AuthenticatedLayout"
 export { default as AdminLayout } from "./AdminLayout"
-export { default as TeamlessLayout } from "./TeamlessLayout"
